+++ src/activities/settings/LanguageSelectActivity.h
@@ -29,6 +29,6 @@ class LanguageSelectActivity final : public Activity {
 
   void onBack() { finish(); }
   ButtonNavigator buttonNavigator;
-  int selectedIndex = 0;
+  uint8_t selectedIndex = 0;
   constexpr static uint8_t totalItems = getLanguageCount();
 };
