+++ src/util/ButtonNavigator.h
@@ -40,11 +40,11 @@ class ButtonNavigator final {
   void onPreviousContinuous(const Callback& callback);
   void onContinuous(const Buttons& buttons, const Callback& callback);
 
-  [[nodiscard]] static int nextIndex(int currentIndex, int totalItems);
-  [[nodiscard]] static int previousIndex(int currentIndex, int totalItems);
+  [[nodiscard]] static uint16_t nextIndex(uint16_t currentIndex, uint16_t totalItems);
+  [[nodiscard]] static uint16_t previousIndex(uint16_t currentIndex, uint16_t totalItems);
 
-  [[nodiscard]] static int nextPageIndex(int currentIndex, int totalItems, int itemsPerPage);
-  [[nodiscard]] static int previousPageIndex(int currentIndex, int totalItems, int itemsPerPage);
+  [[nodiscard]] static uint16_t nextPageIndex(uint16_t currentIndex, uint16_t totalItems, uint16_t itemsPerPage);
+  [[nodiscard]] static uint16_t previousPageIndex(uint16_t currentIndex, uint16_t totalItems, uint16_t itemsPerPage);
 
   [[nodiscard]] static Buttons getNextButtons() {
     return {MappedInputManager::Button::Down, MappedInputManager::Button::Right};
